requirements.txt
python-dateutil==2.8.2
python-dotenv==0.19.2
types-python-dateutil==2.8.19
types-PyYAML==6.0.11
types-requests==2.28.9
requests==2.28.1
web3==5.30.0
pandas~=1.5.0
tqdm~=4.64.1
Jinja2~=3.1.2